backend/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
flask-marshmallow==0.14.0
tldextract==3.1.0